Make ndjsonChunks private to FetchService
XH.fetchNdjson() is the sole public entry point - no compelling case for a standalone Response-level util, and a private helper is easy to promote later if one emerges.
