+++ svc/FetchService.ts
@@ -20,7 +20,6 @@ import {
 } from '@xh/hoist/core';
 import {Exception, HoistException, TimeoutException} from '@xh/hoist/exception';
 import {PromiseTimeoutSpec} from '@xh/hoist/promise';
-import {ndjsonChunks} from '@xh/hoist/utils/async';
 import {isLocalDate, SECONDS} from '@xh/hoist/utils/datetime';
 import {apiDeprecated, warnIf} from '@xh/hoist/utils/js';
 import {StatusCodes} from 'http-status-codes';
@@ -840,4 +839,32 @@ export interface FetchException extends HoistException {
      * @see FetchOptions.autoAbortKey
      */
     isFetchAborted: boolean;
+}
+
+//------------------------
+// Implementation
+//------------------------
+/**
+ * Read an NDJSON Response body incrementally, yielding chunks (arrays) of parsed records as
+ * they arrive off the network. Each line is parsed with native JSON.parse, partial trailing
+ * lines are carried across chunk boundaries, and no more than one network chunk of raw text
+ * is buffered.
+ */
+async function* ndjsonChunks(response: Response): AsyncGenerator<PlainObject[]> {
+    const reader = response.body.getReader(),
+        decoder = new TextDecoder();
+    let buffer = '';
+
+    while (true) {
+        const {done, value} = await reader.read();
+        if (done) break;
+
+        buffer += decoder.decode(value, {stream: true});
+        const lines = buffer.split('\n');
+        buffer = lines.pop(); // retain any partial trailing line for the next chunk
+        if (lines.length) yield lines.filter(Boolean).map(it => JSON.parse(it));
+    }
+
+    buffer += decoder.decode();
+    if (buffer.trim()) yield [JSON.parse(buffer)];
 }
